+++ library/services/notification.service.ts
@@ -1,0 +1,27 @@
+import axios, { AxiosInstance, AxiosResponse } from 'axios';
+import { getBaseURL } from 'lib/utils/getBaseURL.util';
+
+const API = ({ headers = {}, params = {} } = {}): AxiosInstance => {
+  const user = JSON.parse(localStorage.getItem('user') as string) ?? {};
+  const BASE_URL = getBaseURL('APP_NOTIFICATION_API');
+
+  const instance = axios.create({
+    baseURL: BASE_URL,
+    headers: {
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + user.jwt,
+      ...headers,
+    },
+    params,
+  });
+
+  return instance;
+};
+
+const NotificationServices = {
+  readNotification: (id: string): Promise<AxiosResponse> => {
+    return API().put(`/notification/${id}`);
+  },
+};
+
+export default NotificationServices;