+++ vite.config.ts
@@ -14,6 +14,14 @@ export default defineConfig({
     watch: {
       ignored: ['**/dist/**'],
     },
+    // Use proxy for temporary only
+    proxy: {
+      '/api': {
+        target: 'https://sonarqube.qwertysystem.net',
+        changeOrigin: true,
+        rewrite: (path) => path.replace(/^\/api/, '/api'),
+      },
+    },
   },
   assetsInclude: ['**/*.lottie'],
 });